controllers removed, functions moved
