--- Btu Database/WebAppCore/Views/Requirements/Details.cshtml
@@ -1,31 +1,0 @@
-﻿@model WebAppCore.Models.Requirement
-
-@{
-    ViewData["Title"] = "Details";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h2>Details</h2>
-
-<div>
-    <h4>Requirement</h4>
-    <hr />
-    <dl class="dl-horizontal">
-        <dt>
-            @Html.DisplayNameFor(model => model.Description)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Description)
-        </dd>
-        <dt>
-            @Html.DisplayNameFor(model => model.Test)
-        </dt>
-        <dd>
-            @Html.DisplayFor(model => model.Test.TestId)
-        </dd>
-    </dl>
-</div>
-<div>
-    <a asp-action="Edit" asp-route-id="@Model.ReqId">Edit</a> |
-    <a asp-action="Index">Back to List</a>
-</div>
