--- Btu Database/WebAppCore/Views/Requirements/Edit.cshtml
@@ -1,42 +1,0 @@
-﻿@model WebAppCore.Models.Requirement
-
-@{
-    ViewData["Title"] = "Edit";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h2>Edit</h2>
-
-<h4>Requirement</h4>
-<hr />
-<div class="row">
-    <div class="col-md-4">
-        <form asp-action="Edit">
-            <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <input type="hidden" asp-for="ReqId" />
-            <div class="form-group">
-                <label asp-for="TestId" class="control-label"></label>
-                <select asp-for="TestId" class="form-control" asp-items="ViewBag.TestId"></select>
-                <span asp-validation-for="TestId" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="TestVersion" class="control-label"></label>
-                <select asp-for="TestVersion" class="form-control" asp-items="ViewBag.TestVersion"></select>
-                <span asp-validation-for="TestVersion" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <label asp-for="Description" class="control-label"></label>
-                <input asp-for="Description" class="form-control" />
-                <span asp-validation-for="Description" class="text-danger"></span>
-            </div>
-            <div class="form-group">
-                <input type="submit" value="Save" class="btn btn-default" />
-            </div>
-        </form>
-    </div>
-</div>
-
-<div>
-    <a asp-action="Index">Back to List</a>
-</div>
-
