--- Btu Database/WebAppCore/Views/Requirements/Index.cshtml
@@ -1,42 +1,0 @@
-﻿@model IEnumerable<WebAppCore.Models.Requirement>
-
-@{
-    ViewData["Title"] = "Index";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h2>Current Requirements</h2>
-
-<p>
-    <a asp-action="ViewTests\Index">Back</a>
-</p>
-<table class="table">
-    <thead>
-        <tr>
-                <th>
-                    @Html.DisplayNameFor(model => model.Description)
-                </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Test)
-            </th>
-            <th></th>
-        </tr>
-    </thead>
-    <tbody>
-@foreach (var item in Model) {
-        <tr>
-            <td>
-                @Html.DisplayFor(modelItem => item.Description)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Test.TestId)
-            </td>
-            <td>
-                <a asp-action="Edit" asp-route-reqId="@item.ReqId" asp-route-testid="@item.TestId">Edit</a> |
-                <a asp-action="Details" asp-route-reqId="@item.ReqId" asp-route-testid="@item.TestId">Details</a> |
-                <a asp-action="Delete" asp-route-reqId="@item.ReqId" asp-route-testid="@item.TestId">Delete</a>
-            </td>
-        </tr>
-}
-    </tbody>
-</table>
