--- Btu Database/WebAppCore/Views/TestProcs/Index.cshtml
@@ -1,72 +1,0 @@
-﻿@model IEnumerable<WebAppCore.Models.TestProc>
-
-@{
-    ViewData["Title"] = "Index";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h2>Current Procedures</h2>
-
-<p>
-    <a asp-action="ViewTests\Index">Back</a>
-</p>
-<table class="table">
-    <thead>
-        <tr>
-                <th>
-                    @Html.DisplayNameFor(model => model.Parameters)
-                </th>
-                <th>
-                    @Html.DisplayNameFor(model => model.Passed)
-                </th>
-                <th>
-                    @Html.DisplayNameFor(model => model.Order)
-                </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Batch)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Proc)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Req)
-            </th>
-            <th>
-                @Html.DisplayNameFor(model => model.Test)
-            </th>
-            <th></th>
-        </tr>
-    </thead>
-    <tbody>
-@foreach (var item in Model) {
-        <tr>
-            <td>
-                @Html.DisplayFor(modelItem => item.Parameters)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Passed)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Order)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Batch.Status)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Proc.Description)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Req.Description)
-            </td>
-            <td>
-                @Html.DisplayFor(modelItem => item.Test.TestId)
-            </td>
-            <td>
-                @Html.ActionLink("Edit", "Edit", new { id=item.TestId, version=item.TestVersion }) |
-                @Html.ActionLink("Details", "Details", new { id=item.ProcId, testid = item.TestId}) |
-                @Html.ActionLink("Delete", "Delete", new { id=item.ProcId, testid = item.TestId })
-            </td>
-        </tr>
-}
-    </tbody>
-</table>
