+++ Btu Database/WebAppCore/Views/ViewTests/CreateProc.cshtml
@@ -11,7 +11,7 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-action="Create">
+        <form asp-action="CreateProcConfirmed">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <input type="hidden" asp-for="TestId" />
             <input type="hidden" asp-for="BatchId" value="0" />
