+++ Btu Database/WebAppCore/Views/ViewTests/CreateReq.cshtml
@@ -11,7 +11,7 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-action="Create">
+        <form asp-action="CreateReqConfirmed">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group">
                 <label asp-for="ReqId" class="control-label"></label>
