+++ Btu Database/WebAppCore/Views/ViewTests/Edit.cshtml
@@ -109,8 +109,8 @@
             </tbody>
         </table>
         <br /><br />
-        <h4>@Html.ActionLink("Add Requirement", "Create", "Requirements", new { TestId = (int)Model.TestId, TestVersion = (int)Model.TestVersion }, null)</h4>
-        <h4>@Html.ActionLink("Add Procedure", "Create", "TestProcs", new { TestId = (int)Model.TestId, TestVersion = (int)Model.TestVersion }, null)</h4>
+        <h4>@Html.ActionLink("Add Requirement", "CreateReq", new { TestId = (int)Model.TestId, TestVersion = (int)Model.TestVersion }, null)</h4>
+        <h4>@Html.ActionLink("Add Procedure", "CreateProc", new { TestId = (int)Model.TestId, TestVersion = (int)Model.TestVersion }, null)</h4>
 
     </div>
 </div>
